переносит список карточек в пропсы main
