+++ project/src/components/app/app.jsx
@@ -1,13 +1,46 @@
 import React from 'react';
 import Main from '../main/main';
-import PropTypes from 'prop-types';
 
-function App({placesAmount}) {
-  return <Main placesAmount={placesAmount} />;
+const cards = [
+  {
+    price: 120,
+    pictureUrl: 'img/apartment-01.jpg',
+    title: 'Beautiful & luxurious apartment at great location',
+    apartmentType: 'Apartment',
+    rating: 80,
+  },
+  {
+    price: 80,
+    pictureUrl: 'img/room.jpg',
+    title: 'Wood and stone place',
+    apartmentType: 'Private room',
+    rating: 80,
+  },
+  {
+    price: 132,
+    pictureUrl: 'img/apartment-02.jpg',
+    title: 'Canal View Prinsengracht',
+    apartmentType: 'Apartment',
+    rating: 80,
+  },
+  {
+    price: 180,
+    pictureUrl: 'img/apartment-03.jpg',
+    title: 'Nice, cozy, warm big bed apartment',
+    apartmentType: 'Apartment',
+    rating: 100,
+  },
+  {
+    price: 80,
+    pictureUrl: 'img/room.jpg',
+    title: 'Wood and stone place',
+    apartmentType: 'Private room',
+    rating: 80,
+  },
+];
+
+function App() {
+  return <Main cards={cards} />;
 }
-
-App.propTypes = {
-  placesAmount: PropTypes.number.isRequired,
-};
 
 export default App;
